fix: use loadModule instead of manually calculating the package.json path
This also fixes support for monorepo.
(TODO: tests)
